server/requirements.txt
Flask==1.1.1
fake-rpi==0.6.3
importlib-metadata==1.5.0
RPi.Gpio==0.7.0